+++ client/src/App.js
@@ -7,20 +7,13 @@ import Reg from '../src/screens/student/Registration'
 
 import OTPver from './componant/OTPverify'
 import ReportMarks from './screens/supervisor/ReportMarks';
-import Addrecord from './screens/project-member/add-record-page/AddRecordPage';
 import StudentSignUpForm from './screens/authentication/signupFormstudent';
 import StaffSignUpForm from './screens/authentication/signupFormstaffmembrt';
 import DisplayRMarks from './screens/supervisor/DisplayRMarks';
-import PresentationSheduleTable from './screens/project-member/presentation-shedule-tab-page/PresentationSheduletabPage';
-import UpdatePresentationShedule from './screens/project-member/update-presetation-shedule-page/UpdatepresetationShedulePage';
 import PrsentationMarks from './screens/examiner/prsentationMarks';
 import PresentationsTable from './screens/examiner/DisplayPMarks'
 import SidebarExaminer from './componant/SidebarExaminer';
 import SidebarProjMember from './componant/project-member/side-bar-proj-member/SidebarProjMember';
-import Addpresentationpannel from './screens/project-member/add-presentation-pannel-page/PresentationPannelPage'
-import Updatepresentationshedule from './screens/project-member/update-presentation-pannel-page/UpdatePresentationPannelPage'
-
-import PresentationPannelTablep from './screens/project-member/presentation-pannel-tab-page/PresentationPanneltabPage'
 import AssignProjectForm from './screens/coordinator/assignProjectMember';
 import AddAssesment from './screens/coordinator/addAssesment';
 import LoginForm from './screens/authentication/login';
@@ -77,13 +70,7 @@ function App() {
             <Route path='/edit' element={<Edit/>}/>
 
           {/* Rusith */}
-            <Route path='/presentation-shedule' element={<PresentationSheduleTable />} />
-            <Route path='/presentation-shedule/update/:id' element={<UpdatePresentationShedule />} />
             <Route path='/projMemberSideBar' element={<SidebarProjMember />} />
-            <Route path='/presentation-pannel/addpresentationpannel' element={<Addpresentationpannel />} />
-            <Route path='/presentation-pannel/update/:id' element={<Updatepresentationshedule/>} />
-            <Route path='/presentation-pannel' element={<PresentationPannelTablep/>} />
-            <Route path='/presentation-shedule/addrecord' element={<Addrecord />} />
               
           {/* Supun */}
             <Route path='/Registration' element={<Reg />} />
