--- client/src/componant/project-member/presentation-pannel-tab-comp/presentationPanneltabComp.scss
@@ -1,253 +1,0 @@
-
-// .Russa_PresentationPannelTable {
-//   h1 {
-//     color: #003049;
-//     font-size: 25px;
-//     margin-left: 10px;
-//     padding: 10px;
-//     text-align: center;
-//     text-shadow: #ffffff;
-//     font-family: sans-serif;
-//     font-weight: 700;
-//     padding-top: 20px;
-//   }
-// }
-
-
-// .Russa_PresentationPannelTable_searchContainer {
-//   display: flex;
-//   align-items: center;
-//   justify-content: space-between;
-//   margin-bottom: 20px;
-// }
-
-// .Russa_PresentationPannelTable_export,
-// .Russa_PresentationPannelTable_addNew {
-//   background-color: #ffb703;
-//   color: #003049;
-//   font-size: 16px;
-//   font-weight: 500;
-//   border: none;
-//   padding: 8px 12px;
-//   border-radius: 5px;
-//   cursor: pointer;
-//   transition: background-color 0.2s;
-// }
-
-// .Russa_PresentationPannelTable_export:hover,
-// .Russa_PresentationPannelTable_addNew:hover {
-//   background-color: #ffcb57;
-// }
-
-// .Russa_PresentationPannelTable_input_search {
-//   display: flex;
-//   align-items: center;
-//   width: auto; /* Adjust as needed */
-// }
-
-// .Russa_PresentationPannelTable_searchIcon {
-//   color: #ffb43c;
-//   margin-right: 5px;
-// }
-
-// .Russa_PresentationPannelTable_input_field_search {
-//   border: none; /* Removed border */
-//   border-bottom: solid 2px #ffb43c; /* Added bottom border */
-//   border-radius: 0; /* Remove border radius */
-//   padding: 5px;
-//   width: 200px; /* Adjust width as needed */
-//   background-color: transparent; /* Set background color to transparent */
-//   outline: none; /* Remove outline */
-// }
-
-// .Russa_PresentationPannelTable_presentationGridContainer {
-//   border: 1px solid #003049;
-//   border-radius: 5px;
-//   padding: 20px;
-//   box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-//   margin-top: 30px;
-//   background-color: #d8d6d6;
-// }
-
-// .Russa_PresentationPannelTable_cellAction {
-//   display: flex;
-//   align-items: center;
-//   gap: 10px;
-  
-// }
-
-// .Russa_PresentationPannelTable_deleteButton,
-// .Russa_PresentationPannelTable_updateButton {
-//   padding: 8px;
-//   border: none !important;
-//   border-radius: 50%;
-//   color: #ffffff;
-//   cursor: pointer;
-//   display: flex;
-//   align-items: center;
-//   justify-content: center;
-//   transition: background-color 0.3s;
-// }
-
-// .Russa_PresentationPannelTable_deleteButton svg,
-// .Russa_PresentationPannelTable_updateButton svg {
-//   font-size: 18px;
-// }
-
-// .Russa_PresentationPannelTable_deleteButton {
-//   background-color: #dc3545;
-// }
-
-// .Russa_PresentationPannelTable_updateButton {
-//   background-color: #007bff;
-// }
-
-// .Russa_PresentationPannelTable_deleteButton:hover {
-//   background-color: #c82333;
-// }
-
-// .Russa_PresentationPannelTable_updateButton:hover {
-//   background-color: #0056b3;
-// }
-
-// @media screen and (max-width: 768px) {
-//   .Russa_PresentationPannelTable_export,
-//   .Russa_PresentationPannelTable_addNew {
-//     padding: 6px 10px;
-//     font-size: 14px;
-//   }
-
-//   .Russa_PresentationPannelTable_input_search {
-//     max-width: 200px;
-//   }
-// }
-
-
-
-/* General Styling */
-.Russa_PresentationPannelTable {
-  font-family: Arial, sans-serif;
-}
-
-/* Heading Styles */
-.Russa_PresentationPannelTable h1 {
-  color: #003049;
-  font-size: 25px;
-  font-weight: bold;
-  text-align: center;
-  margin-bottom: 35px;
-  padding: 20px 0;
-}
-
-/* Search Container */
-.Russa_PresentationPannelTable_searchContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 20px;
-}
-
-/* Export and Add New Buttons */
-.Russa_PresentationPannelTable_export,
-.Russa_PresentationPannelTable_addNew {
-  background-color: #ffb703;
-  color: #003049;
-  font-size: 16px;
-  font-weight: bold;
-  border: none;
-  padding: 8px 12px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.Russa_PresentationPannelTable_export:hover,
-.Russa_PresentationPannelTable_addNew:hover {
-  background-color: #ffcb57;
-}
-
-/* Search Input Field */
-.Russa_PresentationPannelTable_input_search {
-  display: flex;
-  align-items: center;
-}
-
-.Russa_PresentationPannelTable_searchIcon {
-  color: #ffb43c;
-  margin-right: 5px;
-}
-
-.Russa_PresentationPannelTable_input_field_search {
-  border: none;
-  border-bottom: 2px solid #ffb43c;
-  padding: 5px;
-  width: 200px;
-  background-color: transparent;
-  outline: none;
-}
-
-/* Presentation Grid Container */
-.Russa_PresentationPannelTable_presentationGridContainer {
-  border: 1px solid #003049;
-  border-radius: 5px;
-  padding: 20px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  margin-top: 30px;
-  // background-color: #d8d6d6;
-}
-
-/* Cell Action Buttons */
-.Russa_PresentationPannelTable_cellAction {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.Russa_PresentationPannelTable_deleteButton,
-.Russa_PresentationPannelTable_updateButton {
-  padding: 8px;
-  border: none;
-  border-radius: 50%;
-  color: #ffffff;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: background-color 0.3s;
-}
-
-.Russa_PresentationPannelTable_deleteButton svg,
-.Russa_PresentationPannelTable_updateButton svg {
-  font-size: 18px;
-}
-
-/* Delete Button */
-.Russa_PresentationPannelTable_deleteButton {
-  background-color: #dc3545;
-}
-
-.Russa_PresentationPannelTable_deleteButton:hover {
-  background-color: #c82333;
-}
-
-/* Update Button */
-.Russa_PresentationPannelTable_updateButton {
-  background-color: #007bff;
-}
-
-.Russa_PresentationPannelTable_updateButton:hover {
-  background-color: #0056b3;
-}
-
-/* Responsive Design */
-@media screen and (max-width: 768px) {
-  .Russa_PresentationPannelTable_export,
-  .Russa_PresentationPannelTable_addNew {
-    padding: 6px 10px;
-    font-size: 14px;
-  }
-
-  .Russa_PresentationPannelTable_input_search {
-    max-width: 200px;
-  }
-}
