+++ client/src/componant/project-member/presentation-shedule-tab-comp/presentationSheduletabComp.scss
@@ -61,7 +61,7 @@
     outline: none; /* Remove outline */
   }
   
-  .Russa_PresentationShedeuleTable_presentationGridContainer {
+  .Russa_PresentationSheduleTable_presentationGridContainer {
     border: 1px solid #003049;
     border-radius: 5px;
     padding: 20px;
@@ -69,6 +69,7 @@
     margin-top: 30px;
   }
   
+
   .Russa_PresentationShedeuleTable_cellAction {
     display: flex;
     align-items: center;
